app/requirements.txt
flask==2.3.3
gunicorn==21.2.0
joblib==1.3.2
numpy==1.24.4
pandas==2.0.3
scikit-learn==1.3.0
google-cloud-bigquery==3.11.4
prometheus-client==0.17.1
scipy==1.10.1
db-dtypes==1.1.1